requirements.txt
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
geographiclib==1.52
geopy==2.2.0
idna==3.4
multidict==6.0.2
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
yarl==1.8.1
